widget_demo: the test is test_demo, a graphics window like testg
The demonstrator's test program moves to test_demo.c at the top level,
built as testg is: a graphics window with the demonstration widget
centered in it, for examination and test. The kick count writes under
the widget, and a window resize kills and recreates the widget at the
new center, exercising the create/kill lifecycle. Run from the
repository root, where the widget finds its frames and sound.
